src/requirements.txt
# Flask
flask==1.1.1

# JSON Web Token lib
pyjwt==1.7.1

# Timezone format
pytz==2020.1

# Gunicorn webserver
gunicorn==20.0.4

requests==2.24.0